services/event_store/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
prometheus-fastapi-instrumentator
kafka-python